added initial test and moved test folder
