+++ .github/workflows/ci.yml
@@ -18,13 +18,13 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@v4
 
-    # 2. Set up Python (adjust version as needed)
+    # 2. Set up Python
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
         python-version: "3.13"
 
-    # 3. Install dependencies (using uv since you use it)
+    # 3. Install dependencies 
     - name: Install dependencies with uv
       run: |
         curl -LsSf https://astral.sh/uv/install.sh | sh
